+++ flutter/test/mocks.dart
@@ -1,6 +1,7 @@
 import 'package:mockito/annotations.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/platform/platform.dart';
+import 'package:sentry/src/platform_checker.dart';
 
 const fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 
@@ -62,4 +63,98 @@ class MockPlatform implements Platform {
 
   @override
   bool get isFuchsia => (operatingSystem == 'fuchsia');
+}
+
+class MockPlatformChecker implements PlatformChecker {
+  MockPlatformChecker({
+    this.isDebug = false,
+    this.isProfile = false,
+    this.isRelease = false,
+    this.isWebValue = false,
+    this.hasNativeIntegration = false,
+    Platform? mockPlatform,
+  }) : _mockPlatform = mockPlatform ?? MockPlatform();
+
+  final bool isDebug;
+  final bool isProfile;
+  final bool isRelease;
+  final bool isWebValue;
+  late final Platform _mockPlatform;
+
+  @override
+  bool hasNativeIntegration = false;
+
+  @override
+  bool isDebugMode() => isDebug;
+
+  @override
+  bool isProfileMode() => isProfile;
+
+  @override
+  bool isReleaseMode() => isRelease;
+
+  @override
+  bool get isWeb => isWebValue;
+
+  @override
+  Platform get platform => _mockPlatform;
+}
+
+// Does nothing or returns default values.
+// Usefull for when a Hub needs to be passed but is not used.
+class NoOpHub implements Hub {
+  @override
+  void addBreadcrumb(Breadcrumb crumb, {hint}) {}
+
+  @override
+  void bindClient(SentryClient client) {}
+
+  @override
+  Future<SentryId> captureEvent(
+    SentryEvent event, {
+    stackTrace,
+    hint,
+    ScopeCallback? withScope,
+  }) async {
+    return SentryId.empty();
+  }
+
+  @override
+  Future<SentryId> captureException(
+    throwable, {
+    stackTrace,
+    hint,
+    ScopeCallback? withScope,
+  }) async {
+    return SentryId.empty();
+  }
+
+  @override
+  Future<SentryId> captureMessage(
+    String? message, {
+    SentryLevel? level,
+    String? template,
+    List? params,
+    hint,
+    ScopeCallback? withScope,
+  }) async {
+    return SentryId.empty();
+  }
+
+  @override
+  Hub clone() {
+    return NoOpHub();
+  }
+
+  @override
+  Future<void> close() async {}
+
+  @override
+  void configureScope(ScopeCallback callback) {}
+
+  @override
+  bool get isEnabled => false;
+
+  @override
+  SentryId get lastEventId => SentryId.empty();
 }